xn--d1abkig.online/supermediconline/requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.5.0
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
black==23.7.0
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.5
colorama==0.4.6
Deprecated==1.2.14
Django==4.1.10
frozenlist==1.4.0
greenlet==2.0.2
httplib2==0.22.0
idna==3.4
isort==5.12.0
magic-filter==1.0.9
multidict==6.0.4
mypy-extensions==1.0.0
mysqlclient==2.2.0
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.1
Pillow==10.0.0
platformdirs==3.8.1
protobuf==4.21.12
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.11
pydantic_core==2.1.2
pyparsing==3.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
wrapt==1.15.0
yarl==1.9.2
djangorestframework==3.14.0
django-cors-headers==4.3.0